ci/cd:Controller api통일 및 통합(#83)
