+++ src/main/java/com/archisemtle/semtlewebserverspring/config/SecurityConfig.java
@@ -40,13 +40,14 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
+            .cors(cors -> cors.configurationSource(corsConfigurationSource()))
             .csrf(AbstractHttpConfigurer::disable)
             .sessionManagement(session -> session.sessionCreationPolicy(
                 SessionCreationPolicy.STATELESS)) // Stateless 설정
             .authorizeHttpRequests(auth -> auth
 
-                .requestMatchers(HttpMethod.POST, "/members", "/auth/signin").permitAll()
-                .requestMatchers(HttpMethod.GET, "/api/v1/index/**", "/api/v1/projectboard/**","/promotions/**", "api/v1/activity/**").permitAll()
+                .requestMatchers(HttpMethod.POST, "/api/v1/members", "/api/v1/auth/signin").permitAll()
+                .requestMatchers(HttpMethod.GET, "/api/v1/index/**", "/api/v1/projectboard/**","/api/v1/promotions/**", "api/v1/activity/**").permitAll()
 
                 .requestMatchers(
                     "/swagger-ui/**",       // Swagger UI 리소스
