+++ src/main/java/com/archisemtle/semtlewebserverspring/presentation/TestController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController
+@RestController("/api/v1")
 @CrossOrigin(value = "*")
 public class TestController {
